requirements.txt
# Database
psycopg2-binary
sqlalchemy
python-dotenv

# Core Data Stack
pandas
numpy

# Visualization
matplotlib
seaborn

# Machine Learning
scikit-learn

# Jupyter
notebook
ipykernel

# Utilities
requests

# Testing (optional)
# pytest